infographic/requirements.txt
google-genai==0.5.0
google-generativeai==0.8.3
pydantic==2.10.3
python-dotenv==1.0.1
loguru==0.7.2
